+++ Examples/Example/Unit Tests/Helpers/CoachMarkHelperTests.swift
@@ -5,23 +5,32 @@ import XCTest
 @testable import Instructions
 
 class CoachMarkHelperTests: XCTestCase {
+    // MARK: - Properties
+    private let defaultScreenSize = CGRect(x: 0, y: 0, width: 320, height: 480)
 
-    let instructionsRootView = InstructionsRootView(frame: CGRect(x: 0, y: 0, width: 320, height: 480))
-    let flowManager = MockedFlowManager(coachMarksViewController: CoachMarksViewController())
+    private let flowManager = {
+        MockedFlowManager(coachMarksViewController: CoachMarksViewController())
+    }()
 
-    lazy var coachMarkHelper: CoachMarkHelper = {
+    private let cutoutMaker: (CGRect) -> UIBezierPath = { frame in
+        UIBezierPath(rect: frame)
+    }
+
+    private lazy var instructionsRootView = InstructionsRootView(frame: defaultScreenSize)
+
+    private lazy var coachMarkHelper: CoachMarkHelper = {
         return CoachMarkHelper(instructionsRootView: self.instructionsRootView,
                                flowManager: self.flowManager)
     }()
 
-    override func setUp() {
-        super.setUp()
-    }
-
+    // MARK: - Overridden Methods
     override func tearDown() {
         super.tearDown()
+        instructionsRootView.removeFromSuperview()
+        instructionsRootView.subviews.forEach { $0.removeFromSuperview() }
     }
 
+    // MARK: - 'Update Logic' Tests
     func testThatDefaultCoachMarkIsReturnedWhenViewIsNil() {
 
         let coachMark1 = coachMarkHelper.makeCoachMark()
@@ -106,6 +115,136 @@ class CoachMarkHelperTests: XCTestCase {
         XCTAssertFalse(views2.bodyView.nextLabel.isHidden)
     }
 
+    // MARK: - 'Coordinate Conversion' Tests
+    func testThatCoordinatesAreConvertedProperlyWithInstructionsWindow() {
+        let instructionsWindow = InstructionsRootView(frame: defaultScreenSize)
+        let anchorWindow = InstructionsRootView(frame: defaultScreenSize)
+
+        let anchorSuperview = UIView(frame: CGRect(x: 70, y: 90, width: 100, height: 100))
+        let anchorView = UIView(frame: CGRect(x: 30, y: 40, width: 50, height: 50))
+        let anchorPointOfInterest = CGPoint(x: 55, y: 65)
+
+        let updatedAnchorSuperview = UIView(frame: CGRect(x: 90, y: 25, width: 50, height: 50))
+        let updatedFrame = CGRect(x: 20, y: 70, width: 20, height: 300)
+
+        anchorWindow.addSubview(anchorSuperview)
+        anchorWindow.addSubview(updatedAnchorSuperview)
+        anchorSuperview.addSubview(anchorView)
+        instructionsWindow.addSubview(instructionsRootView)
+
+        let coachMark = coachMarkHelper.makeCoachMark(for: anchorView,
+                                                      cutoutPathMaker: cutoutMaker)
+
+        var coachMark2 = coachMarkHelper.makeCoachMark(forFrame: anchorView.frame,
+                                                       pointOfInterest: anchorPointOfInterest,
+                                                       in: anchorSuperview,
+                                                       cutoutPathMaker: cutoutMaker)
+
+        guard let cutoutPath = coachMark.cutoutPath else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        guard let cutoutPath2 = coachMark2.cutoutPath else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        guard let pointOfInterest2 = coachMark2.pointOfInterest else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        XCTAssertEqual(cutoutPath.bounds, CGRect(x: 100, y: 130, width: 50, height: 50))
+        XCTAssertEqual(cutoutPath2.bounds, CGRect(x: 100, y: 130, width: 50, height: 50))
+        XCTAssertEqual(pointOfInterest2, CGPoint(x: 125, y: 155))
+
+        coachMarkHelper.update(coachMark: &coachMark2,
+                               usingFrame: updatedFrame,
+                               pointOfInterest: nil,
+                               superview: updatedAnchorSuperview,
+                               cutoutPathMaker: cutoutMaker)
+
+        guard let updatedCutoutPath2 = coachMark2.cutoutPath else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        guard let updatedPointOfInterest2 = coachMark2.pointOfInterest else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        XCTAssertEqual(updatedCutoutPath2.bounds, CGRect(x: 110, y: 95, width: 20, height: 300))
+        XCTAssertEqual(updatedPointOfInterest2, CGPoint(x: 125, y: 155))
+    }
+
+    func testThatCoordinatesAreConvertedProperlyWithInstructionsWindowOfArbitrarySize() {
+        let instructionsWindow = InstructionsRootView(frame: CGRect(x: 20, y: 20,
+                                                                    width: 220, height: 340))
+
+        let anchorWindow = InstructionsRootView(frame: defaultScreenSize)
+
+        let anchorSuperview = UIView(frame: CGRect(x: 70, y: 90, width: 100, height: 100))
+        let anchorView = UIView(frame: CGRect(x: 30, y: 40, width: 50, height: 50))
+        let anchorPointOfInterest = CGPoint(x: 55, y: 65)
+
+        let updatedAnchorSuperview = UIView(frame: CGRect(x: 90, y: 25, width: 50, height: 50))
+        let updatedFrame = CGRect(x: 20, y: 70, width: 20, height: 300)
+
+        anchorWindow.addSubview(anchorSuperview)
+        anchorWindow.addSubview(updatedAnchorSuperview)
+        anchorSuperview.addSubview(anchorView)
+        instructionsWindow.addSubview(instructionsRootView)
+
+        let coachMark = coachMarkHelper.makeCoachMark(for: anchorView,
+                                                      cutoutPathMaker: cutoutMaker)
+
+        var coachMark2 = coachMarkHelper.makeCoachMark(forFrame: anchorView.frame,
+                                                       pointOfInterest: anchorPointOfInterest,
+                                                       in: anchorSuperview,
+                                                       cutoutPathMaker: cutoutMaker)
+
+        guard let cutoutPath = coachMark.cutoutPath else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        guard let cutoutPath2 = coachMark2.cutoutPath else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        guard let pointOfInterest2 = coachMark2.pointOfInterest else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        XCTAssertEqual(cutoutPath.bounds, CGRect(x: 80, y: 110, width: 50, height: 50))
+        XCTAssertEqual(cutoutPath2.bounds, CGRect(x: 80, y: 110, width: 50, height: 50))
+        XCTAssertEqual(pointOfInterest2, CGPoint(x: 105, y: 135))
+
+        coachMarkHelper.update(coachMark: &coachMark2,
+                               usingFrame: updatedFrame,
+                               pointOfInterest: nil,
+                               superview: updatedAnchorSuperview,
+                               cutoutPathMaker: cutoutMaker)
+
+        guard let updatedCutoutPath2 = coachMark2.cutoutPath else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        guard let updatedPointOfInterest2 = coachMark2.pointOfInterest else {
+            XCTFail("The cutout path is nil.")
+            return
+        }
+
+        XCTAssertEqual(updatedCutoutPath2.bounds, CGRect(x: 90, y: 75, width: 20, height: 300))
+        XCTAssertEqual(updatedPointOfInterest2, CGPoint(x: 105, y: 135))
+    }
+
+    // MARK: - 'Flow' Tests
     func testThatUpdateDidNotOccur() {
         flowManager.isPaused = true
         flowManager.currentCoachMark = CoachMark()