+++ Sources/Instructions/Extra/Default Views/CoachMarkBodyDefaultView.swift
@@ -116,12 +116,20 @@ private extension CoachMarkBodyDefaultView {
         textView.isScrollEnabled = false
         textView.isUserInteractionEnabled = false
 
-        textView.setContentCompressionResistancePriority(UILayoutPriority.required,
-                                                         for: .horizontal)
+        if #available(iOS 15.0, *) {
+            textView.setContentCompressionResistancePriority(UILayoutPriority.defaultHigh - 1,
+                                                             for: .horizontal)
+        } else {
+            textView.setContentCompressionResistancePriority(UILayoutPriority.required,
+                                                             for: .horizontal)
+        }
+
         textView.setContentCompressionResistancePriority(UILayoutPriority.required,
                                                          for: .vertical)
-        textView.setContentHuggingPriority(UILayoutPriority.defaultHigh, for: .horizontal)
-        textView.setContentHuggingPriority(UILayoutPriority.defaultHigh, for: .vertical)
+        textView.setContentHuggingPriority(UILayoutPriority.defaultHigh,
+                                           for: .horizontal)
+        textView.setContentHuggingPriority(UILayoutPriority.defaultHigh,
+                                           for: .vertical)
 
         return textView
     }
@@ -135,10 +143,22 @@ private extension CoachMarkBodyDefaultView {
 
         label.isUserInteractionEnabled = false
 
-        label.setContentCompressionResistancePriority(UILayoutPriority.required, for: .horizontal)
-        label.setContentCompressionResistancePriority(UILayoutPriority.required, for: .vertical)
-        label.setContentHuggingPriority(UILayoutPriority.required, for: .horizontal)
-        label.setContentHuggingPriority(UILayoutPriority.defaultLow, for: .vertical)
+        if #available(iOS 15.0, *) {
+            label.setContentCompressionResistancePriority(UILayoutPriority.defaultHigh,
+                                                          for: .horizontal)
+            label.setContentHuggingPriority(UILayoutPriority.init(765),
+                                            for: .horizontal)
+        } else {
+            label.setContentCompressionResistancePriority(UILayoutPriority.required,
+                                                          for: .horizontal)
+            label.setContentHuggingPriority(UILayoutPriority.defaultLow,
+                                            for: .horizontal)
+        }
+
+        label.setContentCompressionResistancePriority(UILayoutPriority.required,
+                                                      for: .vertical)
+        label.setContentHuggingPriority(UILayoutPriority.defaultLow,
+                                        for: .vertical)
 
         return label
     }