+++ Sources/Instructions/Helpers/Public/Enums/CoachMarkArrowOrientation.swift
@@ -4,8 +4,8 @@
 import UIKit
 
 /// Available orientations for the arrow.
-/// An arrow can either sit at the top of the body and point upward (.Top) or
-/// sit at the bottom of the body and point downward. (.Bottom)
+/// An arrow can either sit at the top of the body and point upwards (.top) or
+/// sit at the bottom of the body and point downwards. (.bottom)
 public enum CoachMarkArrowOrientation {
     case top
     case bottom